+++ TeacherClient.Avalonia/SplashWindow.axaml.cs
@@ -1,0 +1,13 @@
+using TeacherClient.CrossPlatform.Localization;
+
+namespace TeacherClient.CrossPlatform;
+
+public partial class SplashWindow : Avalonia.Controls.Window
+{
+    public SplashWindow()
+    {
+        InitializeComponent();
+        TitleTextBlock.Text = CrossPlatformText.SplashTitle;
+        SubtitleTextBlock.Text = CrossPlatformText.SplashSubtitle;
+    }
+}